♻️ Final cleanup: extract sanitize module, add confirm controller, fix CSRF
- Extract sanitizeHTML() and initializeSafeHtml() from app.js into a
  dedicated assets/js/sanitize.js module (~50 lines)
- Reduce app.js from ~280 lines to ~7 lines (imports + DOMContentLoaded)
- Create confirm_controller.js Stimulus controller for delete confirmation
  dialogs, replacing inline confirmDelete() JS in the API token page
- Replace dynamically constructed forms with proper <form> elements
  containing server-rendered per-token CSRF tokens
- Add missing CSRF token validation to ApiTokenController::delete()
- Remove superfluous _method=DELETE hidden fields (route accepts POST)

Part of #1045
